+++ tests/testthat/test-full_xampl.R
@@ -1,6 +1,6 @@
 test_that("Full xampl on string", {
   file <- system.file("bib/xampl_mod.bib", package = "bibtex")
-  out <- read.bib(file)
+  expect_message(out <- read.bib(file))
   expect_snapshot(names(out))
 
   expect_snapshot_output(out)
@@ -9,6 +9,6 @@ test_that("Full xampl on string", {
 
 test_that("Full xampl on bibtex", {
   file <- system.file("bib/xampl_mod.bib", package = "bibtex")
-  out <- read.bib(file)
+  expect_message(out <- read.bib(file))
   expect_snapshot_output(toBibtex(out))
 })